Ensure an admin user is removed without issue
Prior to this, if a user was an admin, this error would occur when
attempting to remove them, preventing the Chef Server from converging
successfully:

---- Begin output of chef-server-ctl org-user-remove darksky test ----
STDOUT: User test is in the organization's admin group. Removing users
from an organization without removing them from the admins group is not allowed.
Re-run this command with --force to remove this user from the admins
prior to removing it from the organization.
STDERR: ERROR: Error removing user test from organization darksky.
---- End output of chef-server-ctl org-user-remove darksky test ----